requirements.txt
annotated-types==0.7.0
anyio==4.9.0
aspose-cells-python==25.6.0
aspose-words==25.6.0
beautifulsoup4==4.13.4
black==25.1.0
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
cssutils==2.11.1
currency-symbols==2.0.4
et_xmlfile==2.0.0
fastapi==0.115.6
fonttools==4.58.4
h11==0.16.0
idna==3.10
more-itertools==10.7.0
mypy_extensions==1.1.0
openpyxl==3.1.5
packaging==25.0
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
platformdirs==4.3.8
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyPDF2==3.0.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.41.3
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.20.0
webcolors==24.11.1
xlrd==2.0.2
xls2xlsx==0.2.0
